Added RequestNotificationPermission to OfflineMapAreasApp
