+++ toolkit/offline/src/main/AndroidManifest.xml
@@ -15,12 +15,7 @@
   ~  limitations under the License.
   ~
   -->
-
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <uses-permission android:name="android.permission.INTERNET" />
-    <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
-    <uses-permission android:name="android.permission.FOREGROUND_SERVICE_DATA_SYNC" />
 
     <application>
         <service
